qemu-kvm-ev-2.12.0-44.1.el7_8.1

+++ roms/skiboot/libc/stdio/vsnprintf.c
@@ -26,7 +26,7 @@ static int
 print_str_fill(char **buffer, size_t bufsize, char *sizec,
 					const char *str, char c)
 {
-	int i, sizei, len;
+	size_t i, sizei, len;
 	char *bstart = *buffer;
 
 	sizei = strtoul(sizec, NULL, 10);
@@ -46,7 +46,7 @@ static int
 print_str(char **buffer, size_t bufsize, const char *str)
 {
 	char *bstart = *buffer;
-	int i;
+	size_t i;
 
 	for (i = 0; (i < strlen(str)) && ((*buffer - bstart) < bufsize); i++) {
 		**buffer = str[i];
@@ -164,6 +164,7 @@ print_format(char **buffer, size_t bufsize, const char *format, void *var)
 				break;
 			case 'X':
 				upper = true;
+				/* fallthrough */
 			case 'x':
 				sizec[i] = '\0';
 				value = (unsigned long) var & convert[length_mod];
